Make caching of calculation results optional
For optimising pulse sequences for large superpositions, having the cache is
basically like writing in a memory leak and calling it a feature---the old
results will never be referred to again, so all we're doing is wasting memory.

We still calculate all the values at once to save time, but we only save the
most recent set of angles---this way we get the benefit of the bulk calculation,
but aren't storing unnecessary cruft.