Split search.py up into understandable functions
Before it was just a quickly written mess, but now it can be opened as a module,
and can search specific colour sequences and can be used to target just a single
state, rather than being required to do the whole loop.

Also simplified the logging, so the output printing logic is no longer part of
the calculation code.