Speed up creation of propagators for new angles
Moved to keeping a single copy of the two matrices for a ColourOperator, then
only updating the three diagonals which ever change, and doing it with a single
function call.  This cut down function calls by a massive amount (expensive in
Python), and reduced the complexity of writing the matrices (from O(n^2) to
O(n)).

This scheme also allows us to precalculate all of the square roots necessary at
the initialisation of the ColourOperator class, which cuts down on some of the
longer calculations too (though this probably doesn't matter too much).

This commit gives speedups of around 10x on some test runs for all target states
from (|0> + |1>) to (|1> + |2> + |3> + |4> + |5>).