Improve colour sequence selection rules
The rules imply a minimum and maximum length of pulses, so we impose these
states in the creation.  This should speed up pulse choice for superpositions
with higher maximum states in them (though this will only save a second or two
over the entire run-time, most likely).

More importantly, the maximum length constraint now causes the colour sequences
iterator to terminate cleanly, rather than spinning its way into oblivion as it
filters out every single possibility.  To account for this, the search now has
an error message and termination procedure for running out of available colours.