Make find_sequences.py take command line input
This is a massive overhaul to how the actual human interface script works.
Before, the script file had to be edited to change the search parameters, but
now it accepts targets piped into it, one per line, or it can understand a
couple of commands which will cause it to generate targets itself.

It also just outputs to stdout and stderr, without trying to be clever about it
- we'll just pipe the outputs into the places we want them to go.