Fix "" being an allowable pulse sequence
When targetting the state |0>, the code would try the sequence "" being an
allowable sequence (of course, this is the only acceptable one!).  This would
cause issues with the tests, so rather than specifically hardcode this edge
case, we just disallow 0-length pulse sequences and the code will move on
safely.

This produces a "fatal" message for the 0 case, but if someone actually needs to
optimise that case, they've probably not thought things through.