Fix mathematical error in coloured sidebands
Before, I used
    U_c(x) = exp(-i x [sigma_+ + sigma_-]),        (x = Omega * t)
    U_r(x) = exp(-i x [sigma_+ a + sigma_- a^+]),  (x = ???)
    U_b(x) = exp(-i x [sigma_+ a^+ + sigma_- a]),  (x = ???)
but these use a different phase convention between the carrier and the red and
blue sidebands, so they don't produce correct results.  The angle x is not real
if we use this scheme, which just makes a mess.

This updates the matrices to
    U_c(x) = exp(-i x [sigma_+ + sigma_-]),     (x = Omega * t)
    U_r(x) = exp(x [sigma_+ a - sigma_- a^+]),  (x = eta * Omega * t)
    U_b(x) = exp(x [sigma_+ a^+ - sigma_- a]),  (x = eta * Omega * t)
which are all in the same scheme, so the results make sense.

A side effect of this is that in matrix form, only the carrier brings about
imaginary phases.  This means that some motional superpositions are now
unachievable in the interaction picture if all the phases of the elemnts must be
equal.