Add optimisation of relative phases
Before, I was trying to design a target function which could deal with any
relative phase, but nothing I tried had nice properties.  Functions based on the
absolute difference of each state's occupation from the target value tend to
have sharp edges which the gradient-based optimisation struggles with.
Removing this by squaring the values instead causes there to be large, flat
regions near to local minima which cause precision problems.

This version now takes the relative phases to be an optimisable parameter, so
the minimise method will move the relative phases to match the ideal situation.
The gradients are still available for these, so there's no need to sacrifice on
speed (other than the additional complexity from adding more optimisation
parameters).