Make python terminate with grace period
The Imperial homebrew 'pbsexec' command terminates processes a set time before
the walltime expires.  I do this to the python command so we have time to grab
the data and copy it into somewhere useful.

The command also adds some junk to stdout at the end, but we use stdout for the
python program's output, so I also grab that line and put it in the log file
instead.