Tweak 'pbsexec' rather than trying to fix its mess
The SIGTERM signal message won't necessarily be nicely at the end of the file,
so instead we can grab the 'pbsexec' script and edit it so that it sends its
output to stderr, rather than stdout.  Then we can run our editted version of
it, and we don't have to fiddle the output any more.  I've just stored the
editted version in my home folder.