Fix type-checking on non-floating point numerics
In the `state_specifiers`, setting a target as [0, (1, 1)] should have produced
a target state of (|g0> - |g1>)/sqrt(2), but since the phase was given as an
integer, it would have instead failed the `isinstance(1, float)` check, and the
`state_specifier` would have defaulted to giving a phase of 0.0.

This is fixed by checking if the argument is a `numbers.Number` type instead of
just assuming that `float` will cover it.

This will not have affected any previous results, as the phases are stored in
numpy arrays with `dtype = np.float64`.