Add analytic pulse finding
This is a very quick and dirty hack to get it working at first, and its time
complexity is O(n 2^n) for a target state with n occupied motional levels.

It's very dumb about picking the phases, and it will only return one possible
sequence of colours, even when many exist.  Both of these problems can be solved
by implementing the search as a tree instead, where branches which don't work
are completely pruned.  This allows us to build up all of the possible
combinations using the least possible amount of calculation, though it may be
quite heavy on memory usage.