Make phase selection smarter
When you consider the effect of the pulse sequence evolving forwards in time,
you can see that the phases will always form a checkerboard that looks like
        |g> |e>
    |4>   ...
    |3>  i   1
    |2>  1   i
    |1>  i   1
    |0>  1   i
so we can select the phases in the target based on this.  There's still freedom
to choose the phase up to n*pi, but at the moment the code doesn't attempt to do
this at all in favour of certainly always succeeding.