Simplify analytic.py
This massively simplifies the logic used in analytic.py by grouping all of the
calculations for different colours into one place.  Most of the calculations are
the same regardless of colour.  This also changes the parameters in the
calculations from ground/excited to into/out_of, since most of the logic and
swapping between them is based on this, not on which one is the ground.

This setup should also make it easier to implement a tree to find all of the
pulses, since the algorithm's choices are much easier to change---for the
carrier now we only have to change the `target` parameter to 'e' rather than 'g'
and the logic will propagate it through correctly.