Add binary tree for finding all pulses
This adds a new function `find_all_phases` which returns a binary tree
representing all possible pulse sequences---the tree branches whenever there is
a carrier pulse, because there is a choice of where to put the population.

The binary tree makes calculation of all the possibilities much more efficient
because it does the minimum amount of work necessary---all state that can
possibly be shared is.  It does however store the state of the calculation in
the `Branch` type when it makes one---this is just to make it simpler to trace
the execution of the program from its output, although it is strictly-speaking
unnecessary; all the _required_ information is stored in the `Leaf` nodes.

This also modifies the chequerboard patterning at the beginning to account for
the phases being free up to a factor of pi.  These changes aren't written into
the binary tree, since the calculations will share no state and it won't
simplify the complexity of the calculation.